pyproject.toml
[tool.poetry]
name = "server-config"
version = "0.1.0"
description = ""
authors = ["Mark <mrkswrn@gmail.com>"]
readme = "README.md"
packages = [{include = "./core"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.15.10"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bot = "core.bot:main"
validate = "core.validate:main"